code cleanup - ready for review.
