+++ src/main/scala/scalismo/ui/api/ShowInScene.scala
@@ -23,7 +23,7 @@ import scalismo.image.DiscreteScalarImage
 import scalismo.mesh.{ LineMesh, ScalarMeshField, TriangleMesh, VertexColorMesh3D }
 import scalismo.registration.RigidTransformation
 import scalismo.statisticalmodel.{ DiscreteLowRankGaussianProcess, LowRankGaussianProcess, StatisticalMeshModel, StatisticalVolumeMeshModel }
-import scalismo.tetramesh.{ ScalarVolumeMeshField, TetrahedralMesh, TetrahedralMesh3D }
+import scalismo.tetramesh.{ ScalarVolumeMeshField, TetrahedralMesh }
 import scalismo.ui.model._
 
 import scala.annotation.implicitNotFound
