+++ src/main/scala/scalismo/ui/model/TetrahedralMeshNode.scala
@@ -20,9 +20,8 @@ package scalismo.ui.model
 import java.io.File
 
 import scalismo.geometry.Point3D
-import scalismo.io.{ MeshIO, TetraMeshIO }
-import scalismo.mesh.ScalarMeshField
-import scalismo.tetramesh.{ ScalarVolumeMeshField, TetrahedralMesh3D }
+import scalismo.io.TetraMeshIO
+import scalismo.tetramesh.TetrahedralMesh3D
 import scalismo.ui.model.capabilities._
 import scalismo.ui.model.properties._
 import scalismo.ui.util.{ FileIoMetadata, FileUtil }
